+++ sdk/src/main/java/ly/count/android/sdk/StorageProvider.java
@@ -49,6 +49,8 @@ interface StorageProvider {
 
     String getRemoteConfigValues();//not integrated
 
+    void esWriteCacheToStorage();//required for explicit storage
+
     //fields for data migration
     int getDataSchemaVersion();