+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 22.06.0
   * Fatal crash reports will now be written in sync mode to shared preferences 
+  * Added explicit storage mode where the event queue and request queue are not writen to persistent storage until a signal is given
 
 ## 22.02.3
   * Added CountlyConfigPush object which is used during CountlyPush init.